chore(core): remove npmignore post packaging core
We `build plugin` into `dist/libs/core` after `package core` is ran. This is to include a plugin for
`angular-three` core to provide generators for `angular-three` and the `plugin` can stay a separate
library in the monorepo. However, `ngpackagr` generates a `.npmignore` file to exclude all nested
`package.json`. This breaks the `plugin` for Angular CLI because of ESM/CJS problem. The `plugin`
comes with its own `package.json` with `type:"commonjs"` so that the generators are interpreted as
CJS