feat(core): add NgtObjectEvents directive to pass-through pointer event system
We can use NgtObjectEvents as a directive on elements: ngtObjectEvents. This approach requires the
consumers to pass in a `target` as `Object3D` or `ElementRef<Object3D>` to attach the events on.
Alternatively, we can use this directive as `hostDirective`; this approach can inject the
`NgtObjectEvents` and call `ngtObjectEvents.set()` to set the Object3D value