fix(soba): unwrap `afterNextRender` in `injectAnimations`
BREAKING CHANGE:

This is considered a breaking change because of timing logic has
changed.

`injectAnimations` should NOT control when the `AnimationAction` is
ready or clipped. Previously, this was naively controlled by
`afterNextRender` and the examples have been _unfortunately_ working.

The consumers can migrate to the new behavior with the following
decision making guidance:

- If the object referenced in the `animations` (returned by
`injectGLTF`) are **statically** available on the template (i.e: not
under any control flow or structural directive), then there's no need to
change anything.
- If the object referenced in the `animations` are **dynamically**
available on the template (i.e: under control flow or structural
directive; usually an `*args`), then the `object` parameter should ONLY
be resolved when this object is ready/available. Alternatively, if any
other timing function can be used to ensure the animation object is
ready (i.e: `afterNextRender`), then that can be used instead.