fix(cannon): unwrap `afterNextRender` from `injectBody` and `injectConstraint`
BREAKING CHANGE: This is considered a breaking change because of change
in timing. The CIFs no longer handle timing internally via
`afterNextRender`.

By default, `NgtcPhysics` already handles timing with `afterNextRender`
on the component level (i.e: `NgtcPhysics` component) and these cannon
CIFs are supposed to be used as content children of `NgtcPhysics`.

Additional control is up to the consumers' discretion.