fix(core): assign intermediate ref to the parent store for non-three instance during applyProps
For default objects like the default `MeshBasicMaterial` on a
`ngt-mesh`, we want to still have access to the color space information
on the renderer which is only available on `NgtStore`. In order to do
this, we assign an intermediate reference to the store from the `parent`
during `applyProps` so the instance (`MeshBasicMaterial` in this
example) has access to the Store.

<ngt-mesh>
  <ngt-value [rawValue]="texture()" attach="material.map" />
</ngt-mesh>

things like this. Consumers probably do not run into this because they
would use `ngt-mesh-basic-material#map` instead.