fix(core): add geometryStamp to ngt instance
A new `geometryStamp` is added to `ngt.instanceStore`. This allows the
renderer to update the Local State when a geometry is applied via
`setProperty`. When this is the case, normal `nonObjects()` will not
work because the geometry isn't added via content of the parent.

However, we still need to track / rerun effects on geometries' changes
most of the times. This allows us to notify the Local State when the
geometry property is updated (via `[geometry]`)

A new `nonObjectsChanged` computed is a combination of `nonObjects()`
and `geometryStamp()`. Consumers using `localState.nonObjects()` will
automatically get the changes from `geometryStamp` as well.