fix(soba): clean up a lot of afterNextRender / autoEffect
- Remove a lot of afterNextRender and autoEffect usages. This is due to
misunderstanding of when inputs are resolved. The initial train of
thought was to use afterNextRender to ensure Inputs are resolved in
effects. However, this is not necessarily true because it depends on
when the signals (signals, and computed) are read. effect() is mostly a
safe place to start reacting to inputs; the template itself is an effect
so as long as the result of CIFs are used on the template; we are safe
with the timing of inputs
- code is made more consistent with register deps up front and return
early if possible.